OPIK-395 pr comments
